refactor(models): split each model as independent files.
