+++ bennuhp/admin.py
@@ -1,5 +1,8 @@
 from django.contrib import admin
-from .models import Music, Movie, LiveSchedule
+
+from bennuhp.models.music import Music
+from bennuhp.models.movie import Movie
+from bennuhp.models.liveschedule import LiveSchedule
 
 # To handle in Django administration site
 admin.site.register(Music)
