+++ bennuhp/models/liveschedule.py
@@ -1,0 +1,7 @@
+import uuid
+
+from django.db import models
+
+
+class LiveSchedule(models.Model):
+    uuid = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
